// PlaybookAgent/src/generated/models/Jw_threadcasesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jw_threadcasesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jw_threadcasesstatecode = keyof typeof Jw_threadcasesstatecode;
export const Jw_threadcasesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jw_threadcasesstatuscode = keyof typeof Jw_threadcasesstatuscode;

export interface Jw_threadcasesBase {
  importsequencenumber?: string;
  "jw_caseid@odata.bind": string;
  jw_data?: string;
  jw_name?: string;
  jw_threadcaseid: string;
  "jw_threadid@odata.bind": string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jw_threadcasesstatecode;
  statuscode?: Jw_threadcasesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jw_threadcases extends Jw_threadcasesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  jw_caseidname?: string;
  jw_threadidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  jw_caseid?: object;
  _jw_caseid_value?: string;
  jw_threadid?: object;
  _jw_threadid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}